app/requirements.txt
tortoise-orm
asyncpg
requests
schedule